Move getting descriptor to `TypeName`
