+++ base/src/main/java/io/spine/type/TypeUrl.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017, TeamDev Ltd. All rights reserved.
+ * Copyright 2018, TeamDev Ltd. All rights reserved.
  *
  * Redistribution and use in source and/or binary forms, with or without
  * modification, must retain the above copyright notice and the following
@@ -196,7 +196,7 @@ public final class TypeUrl implements Serializable {
      */
     @Internal
     public GenericDescriptor getDescriptor() {
-        return KnownTypes.getDescriptor(typeName);
+        return TypeName.getDescriptor(typeName);
     }
 
     /**
