+++ base/src/test/java/io/spine/type/KnownTypesShould.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017, TeamDev Ltd. All rights reserved.
+ * Copyright 2018, TeamDev Ltd. All rights reserved.
  *
  * Redistribution and use in source and/or binary forms, with or without
  * modification, must retain the above copyright notice and the following
@@ -21,7 +21,6 @@
 package io.spine.type;
 
 import com.google.protobuf.Any;
-import com.google.protobuf.Descriptors.Descriptor;
 import com.google.protobuf.Duration;
 import com.google.protobuf.Empty;
 import com.google.protobuf.Message;
@@ -42,7 +41,6 @@ import java.util.Set;
 import static io.spine.test.Tests.assertHasPrivateParameterlessCtor;
 import static io.spine.test.Verify.assertSize;
 import static io.spine.type.KnownTypes.getAllFromPackage;
-import static io.spine.type.KnownTypes.getDescriptor;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
@@ -142,20 +140,6 @@ public class KnownTypesShould {
 
         final Collection<TypeUrl> packageTypes = getAllFromPackage(prefix);
         assertTrue(packageTypes.isEmpty());
-    }
-
-    @Test
-    public void provide_proto_descriptor_by_type_name() {
-        final String typeName = "spine.test.types.Task";
-        final Descriptor typeDescriptor = (Descriptor) getDescriptor(typeName);
-        assertNotNull(typeDescriptor);
-        assertEquals(typeName, typeDescriptor.getFullName());
-    }
-
-    @Test(expected = IllegalArgumentException.class)
-    public void fail_to_find_invalid_type_descriptor() {
-        final String invalidTypeName = "no.such.package.InvalidType";
-        getDescriptor(invalidTypeName);
     }
 
     @Test(expected = IllegalStateException.class)
